decouple stories using hooks
